[REFACTOR] #293 - FeedPlayers 베이스 파일 분리 및 네이밍 정리
